Adjust to updated equalsverifier
