+++ services/review/src/test/java/com/backbase/goldensample/review/persistence/ReviewEntityTest.java
@@ -4,7 +4,8 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-
+import nl.jqno.equalsverifier.EqualsVerifier;
+import nl.jqno.equalsverifier.Warning;
 import java.util.Collections;
 import org.junit.jupiter.api.Test;
 
@@ -45,6 +46,20 @@ class ReviewEntityTest {
             () -> assertEquals("subject", reviewEntity.getSubject()),
             () -> assertEquals("content", reviewEntity.getContent()),
             () -> assertEquals(4, reviewEntity.getStars()));
+    }
+
+    /**
+     * Validates the equality of this Entity class in different states and with different data. Due to the definition of
+     * our Entity, 2 suppress warnings are required in order to validate it as explained in <a
+     * href="https://jqno.nl/equalsverifier/manual/jpa-entities/">here</a>.
+     */
+    @Test
+    void validateEntityEqualityAmongStates() {
+        EqualsVerifier.forClass(ReviewEntity.class)
+                .suppress(Warning.IDENTICAL_COPY_FOR_VERSIONED_ENTITY)
+                .suppress(Warning.STRICT_HASHCODE)
+                .suppress(Warning.SURROGATE_KEY)
+                .verify();
     }
 
 }
